Rework Gen3SaveManager into a template class
